+++ packages/migrate/migrations/routes/migrate_page_js/index.js
@@ -10,14 +10,19 @@ import {
 	manual_migration,
 	manual_return_migration,
 	parse,
-	rewrite_returns
+	rewrite_returns,
+	rewrite_type,
+	unwrap
 } from '../utils.js';
 import * as TASKS from '../tasks.js';
 
 const give_up = `${error('Update load function', TASKS.PAGE_LOAD)}\n\n`;
 
-/** @param {string} content */
-export function migrate_page(content) {
+/**
+ * @param {string} content
+ * @param {string} filename
+ */
+export function migrate_page(content, filename) {
 	// early out if we can tell there's no load function
 	// without parsing the file
 	if (!/load/.test(content)) return content;
@@ -32,6 +37,9 @@ export function migrate_page(content) {
 		return content;
 	}
 
+	const match = /__layout(?:-([^.@]+))?/.exec(filename);
+	const load_name = match?.[1] ? `LayoutLoad.${match[1]}` : match ? `LayoutLoad` : 'PageLoad';
+
 	for (const statement of file.ast.statements) {
 		const fn = name ? get_function_node(statement, name) : undefined;
 		if (fn?.body) {
@@ -40,8 +48,11 @@ export function migrate_page(content) {
 			/** @type {Set<string>} */
 			const imports = new Set();
 
+			rewrite_type(fn, file.code, 'Load', load_name);
+
 			rewrite_returns(fn.body, (expr, node) => {
-				const nodes = ts.isObjectLiteralExpression(expr) && get_object_nodes(expr);
+				const value = unwrap(expr);
+				const nodes = ts.isObjectLiteralExpression(value) && get_object_nodes(value);
 
 				if (nodes) {
 					const keys = Object.keys(nodes).sort().join(' ');
@@ -50,8 +61,12 @@ export function migrate_page(content) {
 						return; // nothing to do
 					}
 
-					if (keys === 'props') {
-						automigration(expr, file.code, dedent(nodes.props.getText()));
+					if (
+						keys === 'props' ||
+						((keys === 'status' || keys === 'props status') &&
+							Number(nodes.status.getText()) === 200)
+					) {
+						automigration(value, file.code, dedent(nodes.props?.getText() || ''));
 						return;
 					}
 
@@ -74,10 +89,16 @@ export function migrate_page(content) {
 						if (nodes.error) {
 							const message = is_string_like(nodes.error)
 								? nodes.error.getText()
-								: is_new(nodes.error, 'Error') && nodes.error.arguments[0].getText();
+								: is_new(nodes.error, 'Error')
+								? /** @type {string | undefined} */ (nodes.error.arguments[0]?.getText())
+								: false;
 
-							if (message) {
-								automigration(node, file.code, `throw error(${status || 500}, ${message});`);
+							if (message !== false) {
+								automigration(
+									node,
+									file.code,
+									`throw error(${status || 500}${message ? `, ${message}` : ''});`
+								);
 								imports.add('error');
 								return;
 							}
@@ -102,6 +123,18 @@ export function migrate_page(content) {
 			}
 
 			return file.code.toString();
+		}
+
+		if (ts.isImportDeclaration(statement) && statement.importClause) {
+			const bindings = statement.importClause.namedBindings;
+
+			if (bindings && !ts.isNamespaceImport(bindings)) {
+				for (const binding of bindings.elements) {
+					if (binding.name.escapedText === 'Load') {
+						file.code.overwrite(binding.getStart(), binding.getEnd(), load_name);
+					}
+				}
+			}
 		}
 	}