+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,45 @@
 # create-svelte
 
+## 2.0.0-next.158
+
+### Patch Changes
+
+- Add notes about includes/excludes and path aliases ([#6085](https://github.com/sveltejs/kit/pull/6085))
+
+* setup prettier plugin-search-dirs to enable use with pnpm ([#6101](https://github.com/sveltejs/kit/pull/6101))
+
+## 2.0.0-next.157
+
+### Patch Changes
+
+- Run svelte-kit sync before svelte-check in check scripts ([#6037](https://github.com/sveltejs/kit/pull/6037))
+
+## 2.0.0-next.156
+
+### Patch Changes
+
+- update `app.d.ts` for library skeleton template ([#6020](https://github.com/sveltejs/kit/pull/6020))
+
+## 2.0.0-next.155
+
+### Patch Changes
+
+- make variable names more descriptive ([#5983](https://github.com/sveltejs/kit/pull/5983))
+
+## 2.0.0-next.154
+
+### Patch Changes
+
+- remove session remnants ([#5966](https://github.com/sveltejs/kit/pull/5966))
+
+* remove outdated notes in `api.ts` ([#5964](https://github.com/sveltejs/kit/pull/5964))
+
+## 2.0.0-next.153
+
+### Patch Changes
+
+- Update templates ([#5778](https://github.com/sveltejs/kit/pull/5778))
+
 ## 2.0.0-next.152
 
 ### Patch Changes